+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 
 * Added citation support
 * Added named lanes support
+* Allowed inline mbl
+* Added tick rate in the mbl format
+* Allowed repeat strategy on mbl itself
 
 ## 0.1.2.0 -- 2020-08-16